added mux for http, refactor network module
