+++ network/p2p/p2p.go
@@ -1,4 +1,4 @@
-package network
+package p2p
 
 import (
 	"context"
@@ -14,12 +14,13 @@ import (
 	drouting "github.com/libp2p/go-libp2p/p2p/discovery/routing"
 	dutil "github.com/libp2p/go-libp2p/p2p/discovery/util"
 	"github.com/topology-gg/gram/config"
+	ex "github.com/topology-gg/gram/execution"
 )
 
 type P2P struct {
 	ctx       context.Context
 	errCh     chan error
-	mediator  NetworkMediator
+	executor  ex.Execution
 	host      host.Host
 	namespace string
 	maxPeers  int
@@ -34,7 +35,7 @@ type Stream struct {
 	subscription *pubsub.Subscription
 }
 
-func NewP2P(ctx context.Context, errCh chan error, mediator NetworkMediator, cfg *config.P2pConfig) (*P2P, error) {
+func NewP2P(ctx context.Context, errCh chan error, executor ex.Execution, cfg *config.P2pConfig) (*P2P, error) {
 	namespace, maxPeers, port := cfg.Namespace, cfg.MaxPeers, cfg.Port
 
 	listenAddr := fmt.Sprintf("/ip4/0.0.0.0/tcp/%d", port)
@@ -61,7 +62,7 @@ func NewP2P(ctx context.Context, errCh chan error, mediator NetworkMediator, cfg
 	return &P2P{
 		ctx:       ctx,
 		errCh:     errCh,
-		mediator:  mediator,
+		executor:  executor,
 		host:      host,
 		namespace: namespace,
 		maxPeers:  maxPeers,
@@ -210,7 +211,7 @@ func (p2p *P2P) p2pMessageHandler(subscription *pubsub.Subscription) {
 			continue
 		}
 
-		p2p.mediator.MessageHandler(string(message.Message.Data), SourceP2P)
+		p2p.executor.Execute(string(message.Message.Data))
 	}
 }
 
@@ -235,4 +236,8 @@ func (p2p *P2P) Shutdown() error {
 
 	fmt.Println("P2P host successfully shutted down")
 	return nil
+}
+
+func (p2p *P2P) HostId() string {
+	return p2p.host.ID().String()
 }
